Consolidate as much code as possible between bilinear and bicubic move most of the work in the base class refs #20101
