+++ modules/fluid_properties/include/fluidproperties/TabulatedBicubicFluidProperties.h
@@ -105,46 +105,6 @@ protected:
                      unsigned int ncol,
                      const std::vector<Real> & vec,
                      std::vector<std::vector<Real>> & mat);
-
-  /**
-   * If Newton Method jacobian produces NaNs, set variable to min or max depending on situation
-   * @param min_1 minimum value of pressure
-   * @param max_1 maximum value of pressure
-   * @param min_2 minimum value of temperature
-   * @param max_2 maximum value of temperature
-   * @param i track iteration to get correct data point
-   * @param variable_1 pressure value
-   * @param variable_2 temperature value
-   * @param num_nans_1 track number of nans produced for pressure
-   * @param num_nans_2 track number of nans produced for temperature
-   */
-  void checkNaNs(Real min_1,
-                 Real max_1,
-                 Real min_2,
-                 Real max_2,
-                 unsigned int i,
-                 Real & variable_1,
-                 Real & variable_2,
-                 unsigned int & num_nans_1,
-                 unsigned int & num_nans_2);
-  /**
-   * If values go out of user defined range during Newton Method inversion, set variable to min or
-   * max depending on situation
-   * @param min minimum value of variable
-   * @param max maximum value of variable
-   * @param variable variable of interest (pressure or temperature)
-   * @param num_out_bounds track number of values out of user defined bounds
-   */
-  void checkOutofBounds(Real min, Real max, Real & variable, unsigned int & num_out_bounds);
-  // If values go out of user defined range or NaNs are produced during
-  // Newton Method inversion, produce warnings to inform the user
-  void outputWarnings(unsigned int num_nans_p,
-                      unsigned int num_nans_T,
-                      unsigned int num_out_bounds_p,
-                      unsigned int num_out_bounds_T,
-                      unsigned int convergence_failures,
-                      unsigned int number_points,
-                      std::string variable_set);
 };
 
 #pragma GCC diagnostic pop
